+++ src/RepoM.Core.Plugin/RepositoryOrdering/Configuration/IKeyTypeRegistration{T}.cs
@@ -1,34 +1,14 @@
 namespace RepoM.Core.Plugin.RepositoryOrdering.Configuration;
 
 using System;
-using System.Diagnostics;
+using System.Diagnostics.CodeAnalysis;
 
 /// <summary>
 /// Configuration registration per name
 /// </summary>
-// ReSharper disable once UnusedTypeParameter
+[SuppressMessage("Major Code Smell", "S2326:Unused type parameters should be removed", Justification = "Used for registrations in DI container.")]
 public interface IKeyTypeRegistration<T>   
 {
-    public Type ConfigurationType { get; }
-
-    public string Tag { get; }
-}
-
-[DebuggerDisplay($"{{{nameof(Tag)}}}")]
-public sealed class FixedTypeRegistration<T> : IKeyTypeRegistration<T>
-{
-    public FixedTypeRegistration(Type configurationType, string tag)
-    {
-        ConfigurationType = configurationType;
-
-        if (string.IsNullOrEmpty(tag))
-        {
-            throw new ArgumentNullException(nameof(tag));
-        }
-
-        Tag = tag;
-    }
-
     public Type ConfigurationType { get; }
 
     public string Tag { get; }
